Refactor code to implement MVC pattern and add project structure
The commit reorganizes the codebase into a proper MVC architecture with separate models, views, and controllers. It also adds project structure with
