+++ requirements.txt
@@ -1,0 +1,3 @@
+pandas>=2.0.0
+matplotlib>=3.7.0
+PyQt6>=6.4.0
